add interceptor, redirect

+++ src/js/store/api.js
@@ -2,8 +2,8 @@ import React, {useContext, useReducer} from "react";
 import {reducer} from "./reducer";
 import {url1} from "../util/url1";
 import {logUtil} from "../util/log1";
+import {getToken1} from "../security/interceptor";
 import {login1, logout1} from "../service/authService";
-
 
 const GlobalContext = React.createContext()
 
@@ -14,21 +14,27 @@ export const useGlobalContext = () => {
 const initialState = {
     client: {},
     clientFilter: {},
-    error: ""
+    error: "",
+    info1: ''
 }
 
 export const ApiState = ({children}) => {
 
     const [state, dispatch] = useReducer(reducer, initialState);
 
+    const getClient = (filter = {}, history2) => {
+        let token = getToken1(history2)
+        if (!token) {
 
-    const getClient = (filter = {}) => {
+            return
+        }
+
         if (filter.name != null || filter.title != null || filter.phone != null) {
-
             fetch(url1.filter1, {
                 method: 'POST',
                 headers: {
-                    'Content-Type': 'application/json'
+                    'Content-Type': 'application/json',
+                    'Authorization': token
                 },
                 body: JSON.stringify(filter)
             })
@@ -41,9 +47,19 @@ export const ApiState = ({children}) => {
                             })
                             .catch(error => logUtil('getClient- ', error))
                     }
+                    if (response.status === 403) {
+                        history2.push('/login')
+                        return 403
+                    }
                 })
         } else {
-            fetch(url1.get1)
+            fetch(url1.get1,
+                {
+                    headers: {
+                        'Authorization': token
+                    }
+                }
+            )
                 .then(response => {
                     if (response.status === 200) {
                         response.json()
@@ -53,15 +69,26 @@ export const ApiState = ({children}) => {
                             })
                             .catch(error => logUtil('getClient- ', error))
                     }
+
+                    if (response.status === 403) {
+                        history2.push('/login')
+                        return 403
+                    }
                 })
         }
     }
 
-    const updateClient = (client) => {
+    const updateClient = (client, history2) => {
+        let token = getToken1(history2)
+        if (!token) {
+            return
+        }
+
         fetch(url1.update1, {
             method: 'PUT',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': token
             },
             body: JSON.stringify(client)
         })
@@ -74,15 +101,26 @@ export const ApiState = ({children}) => {
                         })
                         .catch(error => logUtil('updateClient- ', error))
                 }
+                if (response.status === 403) {
+                    history2.push('/login')
+                    return 403
+                }
             })
     }
 
-    const addClient = (client) => {
+    const addClient = (client, history2) => {
+
+        let token = getToken1(history2)
+        if (!token) {
+            return
+        }
+
         client.id = null
         fetch(url1.add1, {
             method: 'POST',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': token
             },
             body: JSON.stringify(client)
         })
@@ -99,18 +137,36 @@ export const ApiState = ({children}) => {
                     logUtil('addClient- 400')
                     dispatch({type: 'error', error: "такой номер уже есть"})
                 }
+
+                if (response.status === 403) {
+                    history2.push('/login')
+                    return 403
+                }
             })
     }
 
-    const deleteClient = (id) => {
+    const deleteClient = (id, history2) => {
+
+        let token = getToken1(history2)
+        if (!token) {
+            return
+        }
+
         fetch(url1.delete1 + '/' + id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': token
+            },
         }).then(response => {
                 if (response.ok) {
                     logUtil('deleteClient+ ', response)
                     dispatch({type: 'deleteClient', id})
                 } else {
                     logUtil('deleteClient- ', response)
+                }
+                if (response.status === 403) {
+                    history2.push('/login')
+                    return 403
                 }
             }
         )
@@ -120,8 +176,18 @@ export const ApiState = ({children}) => {
         dispatch({type: 'localFilterClient', filter})
     }
 
-    const loadReportFile = (format) => {
-        fetch(url1.file1 + format)
+    const loadReportFile = (format, history2) => {
+
+        let token = getToken1(history2)
+        if (!token) {
+            return
+        }
+
+        fetch(url1.file1 + format, {
+            headers: {
+                'Authorization': token
+            }
+        })
             .then(response => {
                 logUtil('loadReportFile+ ', response)
                 response.blob().then(blob => {
